add support for containers; add c9s; add podman to roadmap; add email, irc, gh discussion links (#76)
* add support for containers

Add a `"container"` field to the config which specifies which
container to use.  The `"setup"` is also applied to the container,
so it should work for VMs and containers.  Get rid of the
`become: true` for setup - it is not needed for containers, and
sometimes not needed for VMs.  The user should add `--become true`
to the command line if needed.

* add email, irc and gh discussion links

* add podman role to roadmap